chore(supernova): vitifies workers

+++ apps/supernova/src/hooks/useAlertmanagerAPI.js
@@ -11,31 +11,22 @@ import {
   useSilencesLocalItems,
 } from "../components/StoreProvider"
 
-const createWorker = (path) => {
-  return fetch(new URL(path, import.meta.url))
-    .then((r) => r.blob())
-    .then((blob) => {
-      var blobUrl = window.URL.createObjectURL(blob)
-      let worker
+// Create a worker using Vite's way of handling worker files
+const createWorker = (workerPath) => {
+  const worker = new Worker(new URL(workerPath, import.meta.url), {
+    type: "module",
+  })
 
-      const createWorker = () => {
-        if (!worker) worker = new Worker(blobUrl, { type: "module" })
-        return worker
-      }
+  const stopWorker = () => {
+    if (worker) worker.terminate()
+  }
 
-      const stopWorker = () => {
-        if (!worker) return
-        worker.terminate()
-        worker = null
-      }
-
-      return { createWorker, stopWorker }
-    })
+  return { worker, stopWorker }
 }
 
 // create workers
-const alertsWorker = createWorker("workers/alerts.js")
-const silencesWorker = createWorker("workers/silences.js")
+const alertsWorker = createWorker("../workers/alerts.js")
+const silencesWorker = createWorker("../workers/silences.js")
 
 const useAlertmanagerAPI = (apiEndpoint) => {
   const {
@@ -44,7 +35,9 @@ const useAlertmanagerAPI = (apiEndpoint) => {
     setIsUpdating: setAlertsIsUpdating,
     setError: setAlertsError,
   } = useAlertsActions()
+
   const isUserActive = useUserIsActive()
+
   const {
     setSilences,
     setIsUpdating: setSilencesIsUpdating,
@@ -52,160 +45,119 @@ const useAlertmanagerAPI = (apiEndpoint) => {
     setError: setSilencesError,
   } = useSilencesActions()
 
-  //Setup web workers
+  // Setup web workers
   useEffect(() => {
-    let cleanupAlertsWorker
-    let cleanupSilencesWorker
+    let cleanupAlertsWorker = () => alertsWorker.stopWorker()
+    let cleanupSilencesWorker = () => silencesWorker.stopWorker()
 
-    alertsWorker.then(({ createWorker, stopWorker }) => {
-      const worker = createWorker()
-      console.debug("Worker::Setting up ALERTS worker", worker)
-
-      // receive messages from worker
-      worker.onmessage = (e) => {
-        const action = e.data.action
-        switch (action) {
-          case "ALERTS_UPDATE":
-            console.debug("Worker::ALERT_UPDATE::", e.data)
-            setAlertsData({ items: e.data.alerts, counts: e.data.counts })
-            break
-          case "ALERTS_FETCH_START":
-            console.debug("Worker::ALERTS_FETCH_START::")
-            setAlertsIsUpdating(true)
-            break
-          case "ALERTS_FETCH_END":
-            console.debug("Worker::ALERTS_FETCH_END::")
-            setAlertsIsUpdating(false)
-            break
-          case "ALERTS_FETCH_ERROR":
-            console.debug("Worker::ALERTS_FETCH_ERROR::", e.data.error)
-            setAlertsIsUpdating(false)
-            // error comes as object string and have to be parsed
-            setAlertsError(e.data.error)
-            break
-        }
+    // receive messages from worker
+    alertsWorker.worker.onmessage = (e) => {
+      const action = e.data.action
+      switch (action) {
+        case "ALERTS_UPDATE":
+          console.debug("Worker::ALERT_UPDATE::", e.data)
+          setAlertsData({ items: e.data.alerts, counts: e.data.counts })
+          break
+        case "ALERTS_FETCH_START":
+          console.debug("Worker::ALERTS_FETCH_START::")
+          setAlertsIsUpdating(true)
+          break
+        case "ALERTS_FETCH_END":
+          console.debug("Worker::ALERTS_FETCH_END::")
+          setAlertsIsUpdating(false)
+          break
+        case "ALERTS_FETCH_ERROR":
+          console.debug("Worker::ALERTS_FETCH_ERROR::", e.data.error)
+          setAlertsIsUpdating(false)
+          // error comes as object string and have to be parsed
+          setAlertsError(e.data.error)
+          break
       }
+    }
 
-      cleanupAlertsWorker = () => {
-        console.debug("Worker::Terminating Alerts Worker")
-        return stopWorker()
+    // receive messages from worker
+    silencesWorker.worker.onmessage = (e) => {
+      const action = e.data.action
+      switch (action) {
+        case "SILENCES_UPDATE":
+          console.debug("Worker::SILENCES_UPDATE::", e.data)
+          setSilences({
+            items: e.data?.silences,
+            itemsHash: e.data?.silencesHash,
+            itemsByState: e.data?.silencesBySate,
+          })
+          break
+        case "SILENCES_FETCH_START":
+          console.debug("Worker::SILENCES_FETCH_START::")
+          setSilencesIsUpdating(true)
+          break
+        case "SILENCES_FETCH_END":
+          console.debug("Worker::SILENCES_FETCH_END::")
+          setSilencesIsUpdating(false)
+          break
+        case "SILENCES_FETCH_ERROR":
+          console.debug("Worker::SILENCES_FETCH_ERROR::", e.data.error)
+          setSilencesIsUpdating(false)
+          // error comes as object string and have to be parsed
+          setSilencesError(e.data.error)
+          break
       }
-    })
-
-    silencesWorker.then(({ createWorker, stopWorker }) => {
-      const worker = createWorker()
-      console.debug("Worker::Setting up SILENCES worker")
-
-      // receive messages from worker
-      worker.onmessage = (e) => {
-        const action = e.data.action
-        switch (action) {
-          case "SILENCES_UPDATE":
-            console.debug("Worker::SILENCES_UPDATE::", e.data)
-            setSilences({
-              items: e.data?.silences,
-              itemsHash: e.data?.silencesHash,
-              itemsByState: e.data?.silencesBySate,
-            })
-            break
-          case "SILENCES_FETCH_START":
-            console.debug("Worker::SILENCES_FETCH_START::")
-            setSilencesIsUpdating(true)
-            break
-          case "SILENCES_FETCH_END":
-            console.debug("Worker::SILENCES_FETCH_END::")
-            setSilencesIsUpdating(false)
-            break
-          case "SILENCES_FETCH_ERROR":
-            console.debug("Worker::SILENCES_FETCH_ERROR::", e.data.error)
-            setSilencesIsUpdating(false)
-            // error comes as object string and have to be parsed
-            setSilencesError(e.data.error)
-            break
-        }
-      }
-
-      cleanupSilencesWorker = () => {
-        console.debug("Worker::Terminating Silences Worker")
-        return stopWorker()
-      }
-    })
+    }
 
     return () => {
-      cleanupAlertsWorker && cleanupAlertsWorker()
-      cleanupSilencesWorker && cleanupSilencesWorker()
+      cleanupAlertsWorker()
+      cleanupSilencesWorker()
     }
   }, [])
 
-  // Reconfigure the workers each time we get a new endpoint
+  // Reconfigure workers each time we get a new API endpoint
   useEffect(() => {
     if (!apiEndpoint) return
 
-    // set alerts state to loading
     setAlertsIsLoading(true)
-    alertsWorker.then(({ createWorker }) => {
-      const worker = createWorker()
-      // initial config
-      worker.postMessage({
-        action: "ALERTS_CONFIGURE",
-        fetchVars: {
-          apiEndpoint,
-          options: {},
-        },
-        debug: true,
-      })
+    alertsWorker.worker.postMessage({
+      action: "ALERTS_CONFIGURE",
+      fetchVars: { apiEndpoint, options: {} },
+      debug: true,
     })
 
     setSilencesIsLoading(true)
-    silencesWorker.then(({ createWorker }) => {
-      const worker = createWorker()
-      // initial config
-      worker.postMessage({
-        action: "SILENCES_CONFIGURE",
-        apiEndpoint: apiEndpoint,
-      })
+    silencesWorker.worker.postMessage({
+      action: "SILENCES_CONFIGURE",
+      apiEndpoint,
     })
   }, [apiEndpoint])
 
-  // enable/disable watching in the workers
+  // Enable or disable watching in the workers based on user activity
   useEffect(() => {
     if (isUserActive === undefined) return
-    alertsWorker.then(({ createWorker }) => {
-      const worker = createWorker()
-      worker.postMessage({
-        action: "ALERTS_CONFIGURE",
-        watch: isUserActive,
-      })
+
+    alertsWorker.worker.postMessage({
+      action: "ALERTS_CONFIGURE",
+      watch: isUserActive,
     })
-    silencesWorker.then(({ createWorker }) => {
-      const worker = createWorker()
-      worker.postMessage({
-        action: "SILENCES_CONFIGURE",
-        watch: isUserActive,
-      })
+
+    silencesWorker.worker.postMessage({
+      action: "SILENCES_CONFIGURE",
+      watch: isUserActive,
     })
   }, [isUserActive])
 
-  // as soon as we have locally some silences we refetch the them
+  // Handle re-fetching silences when local items change
   const localItems = useSilencesLocalItems()
   useEffect(() => {
-    if (!localItems) return
     // if we have no silences locally we don't need to refetch them otherwise
     // we will end up in an infinite loop
-    if (Object.keys(localItems).length <= 0) return
+    if (!localItems || Object.keys(localItems).length <= 0) return
 
     // Use setTimeout to delay the worker call delayed by 10s
     setTimeout(() => {
-      silencesWorker.then(({ createWorker }) => {
-        const worker = createWorker()
-        worker.postMessage({
-          action: "SILENCES_FETCH",
-        })
-      })
+      silencesWorker.worker.postMessage({ action: "SILENCES_FETCH" })
     }, 10000)
+
     return () => {
-      if (silencesWorker.current) {
-        silencesWorker.current.terminate()
+      if (silencesWorker.worker) {
+        silencesWorker.worker.terminate()
       }
     }
   }, [localItems])
